move association_counter unit tests into own file
